fix(db-tests): repair CI failures from the editorial admin-role bridge
Two real bugs surfaced by CI's database-quality job on PR #144, both in
code this branch touches:

1. news_domain.test.sql's fixture still granted editorial authority via
   the legacy app_private.editorial_memberships row. The new bridge
   migration (20260919130000) makes has_editorial_role() resolve
   exclusively through the Admin staff role model, so that fixture no
   longer granted anything and api.editorial_create_draft() correctly
   started raising news_editorial_forbidden. Switch the fixture to grant
   the content_admin Admin role (editorial.write + .publish +
   .manage_placements), matching the pattern already used by
   news_editorial_admin_bridge.test.sql.

2. news_editorial_admin_bridge.test.sql and news_editorial_lifecycle.test.sql
   called api.editorial_create_draft(..., 2, 'sanitize-html@2.17.5') with a
   bare integer literal for the smallint reading_time_minutes parameter.
   Postgres only offers an assignment cast (not an implicit cast) from
   integer to smallint, so overload resolution failed with "function does
   not exist" once mixed with the call's other untyped string literals.
   Cast the literals explicitly (2::smallint / 5::smallint).

+++ supabase/tests/database/news_editorial_admin_bridge.test.sql
@@ -63,7 +63,7 @@ select extensions.lives_ok(
     'Résumé suffisant pour satisfaire la contrainte de longueur minimale.',
     'markdown', repeat('Contenu de brouillon suffisant. ', 4),
     '<p>Contenu de brouillon suffisant pour le test du pont.</p>',
-    2, 'sanitize-html@2.17.5'
+    2::smallint, 'sanitize-html@2.17.5'
   )$$,
   'an Admin-role editor can create an editorial draft through the bridged has_editorial_role'
 );
@@ -75,7 +75,7 @@ select set_config(
       'Résumé suffisant pour satisfaire la contrainte de longueur minimale.',
       'markdown', repeat('Contenu de brouillon suffisant. ', 4),
       '<p>Contenu de brouillon suffisant pour le test du pont.</p>',
-      2, 'sanitize-html@2.17.5'
+      2::smallint, 'sanitize-html@2.17.5'
     ) ->> 'articleId'
   ),
   true
@@ -100,7 +100,7 @@ select extensions.throws_ok(
     'Résumé suffisant pour satisfaire la contrainte de longueur minimale.',
     'markdown', repeat('Contenu de brouillon suffisant. ', 4),
     '<p>Contenu de brouillon suffisant pour le test du pont.</p>',
-    2, 'sanitize-html@2.17.5'
+    2::smallint, 'sanitize-html@2.17.5'
   )$$,
   '42501', null,
   'a legacy editorial_memberships row with no Admin staff role grants nothing'
